Update images.adoc
Fixed a typo, changed "amd" to "and".  It is a small change but I wanted to see how it flowed.  Also, giving image files a unique name is mentioned a couple of times above, but it is not clear to me the degree of uniqueness required.  Is it just for files in a single images directory or is it broader.  If I want to reuse and image, must it appear in multiple images directories?